Modify container network create for seamless additions
This change adds the ability for a container to have network interfaces
added without needing to restart to get the interfaces online. Adding
the interface with the container online will be faster and allow for
rolling changes in an environment that may effect critical services that
run from within a container.

Change-Id: I68048ae10cdd52fc3b5c43542686e056237a9305
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>